# public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'Space Flight News API'
  version: '0.1'
paths:
  /articles:
    get:
      tags:
        - articles
      summary: 'Retorna uma lista de artigos'
      description: 'Retorna um objeto de artigos'
      operationId: 20f46a4be21edb0179e4df7acaa45c18
      parameters:
        -
          name: paginate
          in: query
          description: 'Número de artigos por paginação'
          required: false
          schema:
            type: integer
            default: 300
            maximum: 1000
            minimum: 1
      responses:
        '200':
          description: 'A list with articles'
          content:
            application/json:
              schema: {  }
    post:
      tags:
        - articles
      summary: 'Cria um novo artigo'
      description: 'Cria um novo artigo'
      operationId: 9df6bd8bd202eed7afab4cf724e19398
      parameters:
        -
          name: featured
          in: query
          required: false
          schema:
            type: boolean
            default: false
        -
          name: title
          in: query
          required: true
          schema:
            type: string
        -
          name: url
          in: query
          required: true
          schema:
            type: string
        -
          name: imageUrl
          in: query
          required: true
          schema:
            type: string
        -
          name: newSite
          in: query
          required: true
          schema:
            type: string
        -
          name: summary
          in: query
          required: false
          schema:
            type: string
        -
          name: publishedAt
          in: query
          required: false
          schema:
            type: string
      responses:
        '201':
          description: OK
          content:
            application/json:
              schema: {  }
        '422':
          description: 'Faltando dados'
  '/articles/{id}':
    get:
      tags:
        - articles
      summary: 'Retorna um artigo pelo ID'
      description: 'Retorna um artigo pelo ID'
      operationId: 90f7485e28819211cd54fc2b866bd33c
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {  }
        '404':
          description: 'Não encontrado'
          content:
            application/json:
              schema: {  }
    put:
      tags:
        - articles
      summary: 'Atualiza um artigo'
      description: 'Atualiza um artigo'
      operationId: 55f5d2f75407b9984d464b45639952da
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
        -
          name: featured
          in: query
          required: false
          schema:
            type: boolean
            default: false
        -
          name: title
          in: query
          required: true
          schema:
            type: string
        -
          name: url
          in: query
          required: true
          schema:
            type: string
        -
          name: imageUrl
          in: query
          required: true
          schema:
            type: string
        -
          name: newSite
          in: query
          required: true
          schema:
            type: string
        -
          name: summary
          in: query
          required: false
          schema:
            type: string
        -
          name: publishedAt
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: {  }
        '422':
          description: 'Faltando dados'
        '404':
          description: 'Não encontrado'
    delete:
      tags:
        - articles
      summary: 'Delete um artigo'
      description: 'Delete um artigo'
      operationId: 7a59e0f1d0499caca995cdd7cceacc70
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: OK
          content:
            application/json:
              schema: {  }
        '404':
          description: 'Não encontrado'